Forward all labels to the snapshotter
This is primarily motivated by the desire to create windows containers
with a rootfs size larger than the snapshotter default.

This can be acheived with the `ctr run` command with the
`--snapshotter-label` argument. The snapshotter filters labels for which
it has an interest and ignores others, so this change simply forwards
all supplied labels instead of making an API change.

For the two cases below the commands run are:

1. `docker run --rm -it mcr.microsoft.com/windows/servercore:ltsc2022 powershell -Command "(gdr C).Free + (gdr C).Used"`
2. `docker run --label "containerd.io/snapshot/windows/rootfs.sizebytes=32212254720" --rm -it mcr.microsoft.com/windows/servercore:ltsc2022 powershell -Command "(gdr C).Free + (gdr C).Used"`

Before this change:
-------------------
Output from 1: `21339549696`
Output from 2: `21339549696`

After this change:
------------------
Output from 1: `21339549696`
Output from 2: `32076963840`